creature: rename CreatureMood
