+++ src/game/box.h
@@ -8,7 +8,6 @@ void TargetBox(LOT_INFO *LOT, int16_t box_number);
 int32_t StalkBox(ITEM_INFO *item, int16_t box_number);
 int32_t EscapeBox(ITEM_INFO *item, int16_t box_number);
 int32_t ValidBox(ITEM_INFO *item, int16_t zone_number, int16_t box_number);
-void CreatureMood(ITEM_INFO *item, AI_INFO *info, int32_t violent);
 int32_t CalculateTarget(PHD_VECTOR *target, ITEM_INFO *item, LOT_INFO *LOT);
 int32_t CreatureCreature(int16_t item_num);
 int32_t BadFloor(
