+++ src/game/creature.h
@@ -4,3 +4,4 @@
 
 void Creature_Initialise(int16_t item_num);
 void Creature_AIInfo(ITEM_INFO *item, AI_INFO *info);
+void Creature_Mood(ITEM_INFO *item, AI_INFO *info, bool violent);
