+++ src/game/objects/creatures/ape.c
@@ -149,7 +149,7 @@ void Ape_Control(int16_t item_num)
             head = info.angle;
         }
 
-        CreatureMood(item, &info, 0);
+        Creature_Mood(item, &info, false);
 
         angle = CreatureTurn(item, ape->maximum_turn);
