+++ src/game/objects/creatures/bear.c
@@ -117,7 +117,7 @@ void Bear_Control(int16_t item_num)
             head = info.angle;
         }
 
-        CreatureMood(item, &info, 1);
+        Creature_Mood(item, &info, true);
 
         angle = CreatureTurn(item, bear->maximum_turn);
