+++ src/game/objects/creatures/cowboy.c
@@ -83,7 +83,7 @@ void Cowboy_Control(int16_t item_num)
             head = info.angle;
         }
 
-        CreatureMood(item, &info, 0);
+        Creature_Mood(item, &info, false);
 
         angle = CreatureTurn(item, cowboy->maximum_turn);
