+++ src/game/objects/creatures/larson.c
@@ -82,7 +82,7 @@ void Larson_Control(int16_t item_num)
             head = info.angle;
         }
 
-        CreatureMood(item, &info, 0);
+        Creature_Mood(item, &info, false);
 
         angle = CreatureTurn(item, person->maximum_turn);
