+++ src/game/objects/creatures/lion.c
@@ -144,7 +144,7 @@ void Lion_Control(int16_t item_num)
             head = info.angle;
         }
 
-        CreatureMood(item, &info, 1);
+        Creature_Mood(item, &info, true);
 
         angle = CreatureTurn(item, lion->maximum_turn);
