+++ src/game/objects/creatures/raptor.c
@@ -89,7 +89,7 @@ void Raptor_Control(int16_t item_num)
             head = info.angle;
         }
 
-        CreatureMood(item, &info, 1);
+        Creature_Mood(item, &info, true);
 
         angle = CreatureTurn(item, raptor->maximum_turn);
