+++ src/game/objects/creatures/rat.c
@@ -96,7 +96,7 @@ void Rat_Control(int16_t item_num)
             head = info.angle;
         }
 
-        CreatureMood(item, &info, 0);
+        Creature_Mood(item, &info, false);
 
         angle = CreatureTurn(item, RAT_RUN_TURN);
 
@@ -232,7 +232,7 @@ void Vole_Control(int16_t item_num)
             head = info.angle;
         }
 
-        CreatureMood(item, &info, 1);
+        Creature_Mood(item, &info, true);
 
         angle = CreatureTurn(item, VOLE_SWIM_TURN);
