+++ src/game/objects/creatures/skate_kid.c
@@ -86,7 +86,7 @@ void SkateKid_Control(int16_t item_num)
             head = info.angle;
         }
 
-        CreatureMood(item, &info, 0);
+        Creature_Mood(item, &info, false);
 
         angle = CreatureTurn(item, SKATE_KID_SKATE_TURN);
