+++ src/game/objects/creatures/torso.c
@@ -92,7 +92,7 @@ void Torso_Control(int16_t item_num)
             head = info.angle;
         }
 
-        CreatureMood(item, &info, 1);
+        Creature_Mood(item, &info, true);
 
         angle =
             Math_Atan(
