+++ src/game/objects/creatures/trex.c
@@ -98,7 +98,7 @@ void TRex_Control(int16_t item_num)
             head = info.angle;
         }
 
-        CreatureMood(item, &info, 1);
+        Creature_Mood(item, &info, true);
 
         angle = CreatureTurn(item, dino->maximum_turn);
